🚚 restoreHistory 파일 위치 변경, history관련 리코일 상태 위치 변경
